requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
discord.py==2.4.0
frozenlist==1.4.1
idna==3.8
multidict==6.0.5
requests==2.32.3
urllib3==2.2.2
yarl==1.9.4
